+++ rdkit/Chem/AtomPairs/Utils.py
@@ -9,6 +9,7 @@
 #
 
 import math
+import warnings
 
 from rdkit import Chem
 from rdkit.Chem import rdMolDescriptors
@@ -106,45 +107,12 @@ GetAtomCode = rdMolDescriptors.GetAtomPairAtomCode
 
 
 def NumPiElectrons(atom):
-  """ Returns the number of electrons an atom is using for pi bonding
+  """ DEPRECATED: please use rdMolDescriptors.GetNumPiElectrons instead.
+  """
 
-    >>> m = Chem.MolFromSmiles('C=C')
-    >>> NumPiElectrons(m.GetAtomWithIdx(0))
-    1
-
-    >>> m = Chem.MolFromSmiles('C#CC')
-    >>> NumPiElectrons(m.GetAtomWithIdx(0))
-    2
-    >>> NumPiElectrons(m.GetAtomWithIdx(1))
-    2
-
-    >>> m = Chem.MolFromSmiles('O=C=CC')
-    >>> NumPiElectrons(m.GetAtomWithIdx(0))
-    1
-    >>> NumPiElectrons(m.GetAtomWithIdx(1))
-    2
-    >>> NumPiElectrons(m.GetAtomWithIdx(2))
-    1
-    >>> NumPiElectrons(m.GetAtomWithIdx(3))
-    0
-
-    >>> m = Chem.MolFromSmiles('c1ccccc1')
-    >>> NumPiElectrons(m.GetAtomWithIdx(0))
-    1
-
-    FIX: this behaves oddly in these cases:
-
-    >>> m = Chem.MolFromSmiles('S(=O)(=O)')
-    >>> NumPiElectrons(m.GetAtomWithIdx(0))
-    2
-
-    >>> m = Chem.MolFromSmiles('S(=O)(=O)(O)O')
-    >>> NumPiElectrons(m.GetAtomWithIdx(0))
-    0
-
-    In the second case, the S atom is tagged as sp3 hybridized.
-
-    """
+  warnings.warn(
+    "The Chem.AtomPairs.Utils.NumPiElectrons is deprecated; please use rdMolDescriptors.GetNumPiElectrons instead.",
+    DeprecationWarning, stacklevel=2)
 
   res = 0
   if atom.GetIsAromatic():